requirements.txt
Deprecated==1.2.15
lxml==5.3.0
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pikepdf==9.4.2
pillow==11.0.0
python-dotenv==1.0.1
typing_extensions==4.12.2
wrapt==1.17.0
